Struct typed_index_collections::TiSlice

#[repr(transparent)]
pub struct TiSlice<K, V> {
    pub raw: [V],
    /* private fields */
}

A dynamically-sized view into a contiguous sequence of T that only accepts keys of the type K.

TiSlice<K, V> is a wrapper around Rust primitive type slice. The struct mirrors the stable API of Rust slice and forwards to it as much as possible.

TiSlice<K, V> uses K instead of usize for element indices. It also uses Range, RangeTo, RangeFrom, RangeInclusive and RangeToInclusive range types with K indices for get-methods and index expressions. The RangeFull trait is not currently supported.

TiSlice<K, V> require the index to implement From<usize> and Into<usize> traits. Their implementation can be easily done with derive_more crate and #[derive(From, Into)].

There are zero-cost conversions available between types and references:

&[V] and &TiSlice<K, V> with AsRef,

&mut [V] and &mut TiSlice<K, V> with AsMut,

Box<[V]> and Box<TiSlice<K, V>> with From and Into.

Added methods:

from_ref - Converts a &[V] into a &TiSlice<K, V>.

from_mut - Converts a &mut [V] into a &mut TiSlice<K, V>.

keys - Returns an iterator over all keys.

next_key - Returns the index of the next slice element to be appended and at the same time number of elements in the slice of type K.

first_key - Returns the first slice element index of type K, or None if the slice is empty.

first_key_value - Returns the first slice element index of type K and the element itself, or None if the slice is empty.

first_key_value_mut - Returns the first slice element index of type K and a mutable reference to the element itself, or None if the slice is empty.

last_key - Returns the last slice element index of type K, or None if the slice is empty.

last_key_value - Returns the last slice element index of type K and the element itself, or None if the slice is empty.

last_key_value_mut - Returns the last slice element index of type K and a mutable reference to the element itself, or None if the slice is empty.

iter_enumerated - Returns an iterator over all key-value pairs. It acts like self.iter().enumerate(), but use K instead of usize for iteration indices.

iter_mut_enumerated - Returns an iterator over all key-value pairs, with mutable references to the values. It acts like self.iter_mut().enumerate(), but use K instead of usize for iteration indices.

position - Searches for an element in an iterator, returning its index of type K. It acts like self.iter().position(...), but instead of usize it returns index of type K.

rposition - Searches for an element in an iterator from the right, returning its index of type K. It acts like self.iter().rposition(...), but instead of usize it returns index of type K.

Example

use typed_index_collections::TiSlice;
use derive_more::{From, Into};

#[derive(From, Into)]
struct FooId(usize);

let mut foos_raw = [1, 2, 5, 8];
let foos: &mut TiSlice<FooId, usize> = TiSlice::from_mut(&mut foos_raw);
foos[FooId(2)] = 4;
assert_eq!(foos[FooId(2)], 4);

Fields
raw: [V]

Raw slice property

Implementations

impl<K, V> TiSlice<K, V>

pub fn from_ref(raw: &[V]) -> &Self

Converts a &[V] into a &TiSlice<K, V>.

Example

pub struct Id(usize);
let slice: &TiSlice<Id, usize> = TiSlice::from_ref(&[1, 2, 4]);

pub fn from_mut(raw: &mut [V]) -> &mut Self

Converts a &mut [V] into a &mut TiSlice<K, V>.

Example

pub struct Id(usize);
let slice: &mut TiSlice<Id, usize> = TiSlice::from_mut(&mut [1, 2, 4]);

pub fn len(&self) -> usize

Returns the number of elements in the slice.

See slice::len for more details.

pub fn next_key(&self) -> Kwhere K: From<usize>,

Returns the index of the next slice element to be appended and at the same time number of elements in the slice of type K.

Example

#[derive(Eq, Debug, From, Into, PartialEq)]
pub struct Id(usize);
let slice: &TiSlice<Id, usize> = TiSlice::from_ref(&[1, 2, 4]);
assert_eq!(slice.next_key(), Id(3));

pub const fn is_empty(&self) -> bool

Returns true if the slice has a length of 0.

See slice::is_empty for more details.

pub fn keys(&self) -> TiSliceKeys<K>where K: From<usize>,

Returns an iterator over all keys.

Example

#[derive(Debug, Eq, From, Into, PartialEq)]
pub struct Id(usize);
let slice: &TiSlice<Id, usize> = TiSlice::from_ref(&[1, 2, 4]);
let mut iterator = slice.keys();
assert_eq!(iterator.next(), Some(Id(0)));
assert_eq!(iterator.next(), Some(Id(1)));
assert_eq!(iterator.next(), Some(Id(2)));
assert_eq!(iterator.next(), None);

pub fn first(&self) -> Option<&V>

Returns the first element of the slice, or None if it is empty.

See slice::first for more details.

pub fn first_mut(&mut self) -> Option<&mut V>

Returns a mutable reference to the first element of the slice, or None if it is empty.

See slice::first_mut for more details.

pub fn first_key(&self) -> Option<K>where K: From<usize>,

Returns the first slice element index of type K, or None if the slice is empty.

Example

#[derive(Debug, Eq, From, Into, PartialEq)]
pub struct Id(usize);
let empty_slice: &TiSlice<Id, usize> = TiSlice::from_ref(&[]);
let slice: &TiSlice<Id, usize> = TiSlice::from_ref(&[1, 2, 4]);
assert_eq!(empty_slice.first_key(), None);
assert_eq!(slice.first_key(), Some(Id(0)));

pub fn first_key_value(&self) -> Option<(K, &V)>where K: From<usize>,

Returns the first slice element index of type K and the element itself, or None if the slice is empty.

See slice::first for more details.

Example

#[derive(Debug, Eq, From, Into, PartialEq)]
pub struct Id(usize);
let empty_slice: &TiSlice<Id, usize> = TiSlice::from_ref(&[]);
let slice: &TiSlice<Id, usize> = TiSlice::from_ref(&[1, 2, 4]);
assert_eq!(empty_slice.first_key_value(), None);
assert_eq!(slice.first_key_value(), Some((Id(0), &1)));

pub fn first_key_value_mut(&mut self) -> Option<(K, &mut V)>where K: From<usize>,

Returns the first slice element index of type K and a mutable reference to the element itself, or None if the slice is empty.

See slice::first_mut for more details.

Example

#[derive(Debug, Eq, From, Into, PartialEq)]
pub struct Id(usize);
let empty_slice: &mut TiSlice<Id, usize> = TiSlice::from_mut(&mut []);
let mut array = [1, 2, 4];
let slice: &mut TiSlice<Id, usize> = TiSlice::from_mut(&mut array);
assert_eq!(empty_slice.first_key_value_mut(), None);
assert_eq!(slice.first_key_value_mut(), Some((Id(0), &mut 1)));
*slice.first_key_value_mut().unwrap().1 = 123;
assert_eq!(slice.raw, [123, 2, 4]);

pub fn split_first(&self) -> Option<(&V, &TiSlice<K, V>)>

Returns the first and all the rest of the elements of the slice, or None if it is empty.

See slice::split_first for more details.

pub fn split_first_mut(&mut self) -> Option<(&mut V, &mut TiSlice<K, V>)>

Returns the first and all the rest of the elements of the slice, or None if it is empty.

See slice::split_first_mut for more details.

pub fn split_last(&self) -> Option<(&V, &TiSlice<K, V>)>

Returns the last and all the rest of the elements of the slice, or None if it is empty.

See slice::split_last for more details.

pub fn split_last_mut(&mut self) -> Option<(&mut V, &mut TiSlice<K, V>)>

Returns the last and all the rest of the elements of the slice, or None if it is empty.

See slice::split_last_mut for more details.

pub fn last(&self) -> Option<&V>

Returns the last element of the slice of type K, or None if it is empty.

See slice::last for more details.

pub fn last_mut(&mut self) -> Option<&mut V>

Returns a mutable reference to the last item in the slice.

See slice::last_mut for more details.

pub fn last_key(&self) -> Option<K>where K: From<usize>,

Returns the last slice element index of type K, or None if the slice is empty.

Example

#[derive(Debug, Eq, From, Into, PartialEq)]
pub struct Id(usize);
let empty_slice: &TiSlice<Id, usize> = TiSlice::from_ref(&[]);
let slice: &TiSlice<Id, usize> = TiSlice::from_ref(&[1, 2, 4]);
assert_eq!(empty_slice.last_key(), None);
assert_eq!(slice.last_key(), Some(Id(2)));

pub fn last_key_value(&self) -> Option<(K, &V)>where K: From<usize>,

Returns the last slice element index of type K and the element itself, or None if the slice is empty.

See slice::last for more details.

Example

#[derive(Debug, Eq, From, Into, PartialEq)]
pub struct Id(usize);
let empty_slice: &TiSlice<Id, usize> = TiSlice::from_ref(&[]);
let slice: &TiSlice<Id, usize> = TiSlice::from_ref(&[1, 2, 4]);
assert_eq!(empty_slice.last_key_value(), None);
assert_eq!(slice.last_key_value(), Some((Id(2), &4)));

pub fn last_key_value_mut(&mut self) -> Option<(K, &mut V)>where K: From<usize>,

Returns the last slice element index of type K and a mutable reference to the element itself, or None if the slice is empty.

See slice::last_mut for more details.

Example

#[derive(Debug, Eq, From, Into, PartialEq)]
pub struct Id(usize);
let empty_slice: &mut TiSlice<Id, usize> = TiSlice::from_mut(&mut []);
let mut array = [1, 2, 4];
let slice: &mut TiSlice<Id, usize> = TiSlice::from_mut(&mut array);
assert_eq!(empty_slice.last_key_value_mut(), None);
assert_eq!(slice.last_key_value_mut(), Some((Id(2), &mut 4)));
*slice.last_key_value_mut().unwrap().1 = 123;
assert_eq!(slice.raw, [1, 2, 123]);

pub fn get<I>(&self, index: I) -> Option<&I::Output>where I: TiSliceIndex<K, V>,

Returns a reference to an element or subslice depending on the type of index or None if the index is out of bounds.

See slice::get for more details.

pub fn get_mut<I>(&mut self, index: I) -> Option<&mut I::Output>where I: TiSliceIndex<K, V>,

Returns a mutable reference to an element or subslice depending on the type of index or None if the index is out of bounds.

See slice::get_mut for more details.

pub unsafe fn get_unchecked<I>(&self, index: I) -> &I::Outputwhere I: TiSliceIndex<K, V>,

Returns a reference to an element or subslice depending on the type of index, without doing bounds checking.

This is generally not recommended, use with caution! Calling this method with an out-of-bounds index is undefined behavior even if the resulting reference is not used. For a safe alternative see get.

See slice::get_unchecked for more details.

pub unsafe fn get_unchecked_mut<I>(&mut self, index: I) -> &mut I::Outputwhere I: TiSliceIndex<K, V>,

Returns a mutable reference to an element or subslice depending on the type of index, without doing bounds checking.

This is generally not recommended, use with caution! Calling this method with an out-of-bounds index is undefined behavior even if the resulting reference is not used. For a safe alternative see get_mut.

See slice::get_unchecked_mut for more details.

pub const fn as_ptr(&self) -> *const V

Returns a raw pointer to the slice’s buffer.

See slice::as_ptr for more details.

pub fn as_mut_ptr(&mut self) -> *mut V

Returns an unsafe mutable reference to the slice’s buffer.

See slice::as_mut_ptr for more details.

pub fn swap(&mut self, a: K, b: K)where usize: From<K>,

Swaps two elements in the slice.

See slice::swap for more details.

pub fn reverse(&mut self)

Reverses the order of elements in the slice, in place.

See slice::reverse for more details.

pub fn iter(&self) -> Iter<'_, V>

Returns an iterator over the slice.

See slice::iter for more details.

pub fn iter_enumerated(&self) -> TiEnumerated<Iter<'_, V>, K, &V>where K: From<usize>,

Returns an iterator over all key-value pairs.

It acts like self.iter().enumerate(), but use K instead of usize for iteration indices.

See slice::iter for more details.

Example

#[derive(Debug, Eq, From, Into, PartialEq)]
pub struct Id(usize);
let slice: &TiSlice<Id, usize> = TiSlice::from_ref(&[1, 2, 4]);
let mut iterator = slice.iter_enumerated();
assert_eq!(iterator.next(), Some((Id(0), &1)));
assert_eq!(iterator.next(), Some((Id(1), &2)));
assert_eq!(iterator.next(), Some((Id(2), &4)));
assert_eq!(iterator.next(), None);

pub fn iter_mut(&mut self) -> IterMut<'_, V>

Returns an iterator that allows modifying each value.

See slice::iter_mut for more details.

pub fn iter_mut_enumerated(&mut self) -> TiEnumerated<IterMut<'_, V>, K, &mut V>where K: From<usize>,

Returns an iterator over all key-value pairs, with mutable references to the values.

It acts like self.iter_mut().enumerate(), but use K instead of usize for iteration indices.

Example

#[derive(Debug, Eq, From, Into, PartialEq)]
pub struct Id(usize);
let mut array = [1, 2, 4];
let slice: &mut TiSlice<Id, usize> = TiSlice::from_mut(&mut array);
for (key, value) in slice.iter_mut_enumerated() {
    *value += key.0;
}
assert_eq!(array, [1, 3, 6]);

pub fn position<P>(&self, predicate: P) -> Option<K>where K: From<usize>, P: FnMut(&V) -> bool,

Searches for an element in an iterator, returning its index of type K.

It acts like self.iter().position(...), but instead of usize it returns index of type K.

See slice::iter and Iterator::position for more details.

Example

#[derive(Debug, Eq, From, Into, PartialEq)]
pub struct Id(usize);
let slice: &TiSlice<Id, usize> = TiSlice::from_ref(&[1, 2, 4, 2, 1]);
assert_eq!(slice.position(|&value| value == 1), Some(Id(0)));
assert_eq!(slice.position(|&value| value == 2), Some(Id(1)));
assert_eq!(slice.position(|&value| value == 3), None);
assert_eq!(slice.position(|&value| value == 4), Some(Id(2)));

pub fn rposition<P>(&self, predicate: P) -> Option<K>where K: From<usize>, P: FnMut(&V) -> bool,

Searches for an element in an iterator from the right, returning its index of type K.

It acts like self.iter().rposition(...), but instead of usize it returns index of type K.

See slice::iter and Iterator::position for more details.

Example

#[derive(Debug, Eq, From, Into, PartialEq)]
pub struct Id(usize);
let slice: &TiSlice<Id, usize> = TiSlice::from_ref(&[1, 2, 4, 2, 1]);
assert_eq!(slice.rposition(|&value| value == 1), Some(Id(4)));
assert_eq!(slice.rposition(|&value| value == 2), Some(Id(3)));
assert_eq!(slice.rposition(|&value| value == 3), None);
assert_eq!(slice.rposition(|&value| value == 4), Some(Id(2)));

pub fn windows(&self, size: usize) -> TiSliceRefMap<Windows<'_, V>, K, V>

Returns an iterator over all contiguous windows of length size. The windows overlap. If the slice is shorter than size, the iterator returns no values.

See slice::windows for more details.

pub fn chunks(&self, chunk_size: usize) -> TiSliceRefMap<Chunks<'_, V>, K, V>

Returns an iterator over chunk_size elements of the slice at a time, starting at the beginning of the slice.

See slice::chunks for more details.

pub fn chunks_mut( &mut self, chunk_size: usize ) -> TiSliceMutMap<ChunksMut<'_, V>, K, V>

Returns an iterator over chunk_size elements of the slice at a time, starting at the beginning of the slice.

See slice::chunks_mut for more details.

pub fn chunks_exact( &self, chunk_size: usize ) -> TiSliceRefMap<ChunksExact<'_, V>, K, V>

Returns an iterator over chunk_size elements of the slice at a time, starting at the beginning of the slice.

See slice::chunks_exact for more details.

pub fn chunks_exact_mut( &mut self, chunk_size: usize ) -> TiSliceMutMap<ChunksExactMut<'_, V>, K, V>

Returns an iterator over chunk_size elements of the slice at a time, starting at the beginning of the slice.

See slice::chunks_exact_mut for more details.

pub fn rchunks(&self, chunk_size: usize) -> TiSliceRefMap<RChunks<'_, V>, K, V>

Returns an iterator over chunk_size elements of the slice at a time, starting at the end of the slice.

See slice::rchunks for more details.

pub fn rchunks_mut( &mut self, chunk_size: usize ) -> TiSliceMutMap<RChunksMut<'_, V>, K, V>

Returns an iterator over chunk_size elements of the slice at a time, starting at the end of the slice.

See slice::rchunks_mut for more details.

pub fn rchunks_exact( &self, chunk_size: usize ) -> TiSliceRefMap<RChunksExact<'_, V>, K, V>

Returns an iterator over chunk_size elements of the slice at a time, starting at the end of the slice.

See slice::rchunks_exact for more details.

pub fn rchunks_exact_mut( &mut self, chunk_size: usize ) -> TiSliceMutMap<RChunksExactMut<'_, V>, K, V>

Returns an iterator over chunk_size elements of the slice at a time, starting at the end of the slice.

See slice::rchunks_exact_mut for more details.

pub fn split_at(&self, mid: K) -> (&Self, &Self)where usize: From<K>,

Divides one slice into two at an index.

See slice::split_at for more details.

pub fn split_at_mut(&mut self, mid: K) -> (&mut Self, &mut Self)where usize: From<K>,

Divides one mutable slice into two at an index.

See slice::split_at_mut for more details.

pub fn split<F>(&self, pred: F) -> TiSliceRefMap<Split<'_, V, F>, K, V>where F: FnMut(&V) -> bool,

Returns an iterator over subslices separated by elements that match pred. The matched element is not contained in the subslices.

See slice::split for more details.

pub fn split_mut<F>( &mut self, pred: F ) -> TiSliceMutMap<SplitMut<'_, V, F>, K, V>where F: FnMut(&V) -> bool,

Returns an iterator over mutable subslices separated by elements that match pred. The matched element is not contained in the subslices.

See slice::split_mut for more details.

pub fn rsplit<F>(&self, pred: F) -> TiSliceRefMap<RSplit<'_, V, F>, K, V>where F: FnMut(&V) -> bool,

Returns an iterator over subslices separated by elements that match pred, starting at the end of the slice and working backwards. The matched element is not contained in the subslices.

See slice::rsplit for more details.

pub fn rsplit_mut<F>( &mut self, pred: F ) -> TiSliceMutMap<RSplitMut<'_, V, F>, K, V>where F: FnMut(&V) -> bool,

Returns an iterator over mutable subslices separated by elements that match pred, starting at the end of the slice and working backwards. The matched element is not contained in the subslices.

See slice::rsplit_mut for more details.

pub fn splitn<F>( &self, n: usize, pred: F ) -> TiSliceRefMap<SplitN<'_, V, F>, K, V>where F: FnMut(&V) -> bool,

Returns an iterator over subslices separated by elements that match pred, limited to returning at most n items. The matched element is not contained in the subslices.

See slice::splitn for more details.

pub fn splitn_mut<F>( &mut self, n: usize, pred: F ) -> TiSliceMutMap<SplitNMut<'_, V, F>, K, V>where F: FnMut(&V) -> bool,

Returns an iterator over subslices separated by elements that match pred, limited to returning at most n items. The matched element is not contained in the subslices.

See slice::splitn_mut for more details.

pub fn rsplitn<F>( &self, n: usize, pred: F ) -> TiSliceRefMap<RSplitN<'_, V, F>, K, V>where F: FnMut(&V) -> bool,

Returns an iterator over subslices separated by elements that match pred limited to returning at most n items. This starts at the end of the slice and works backwards. The matched element is not contained in the subslices.

See slice::rsplitn for more details.

pub fn rsplitn_mut<F>( &mut self, n: usize, pred: F ) -> TiSliceMutMap<RSplitNMut<'_, V, F>, K, V>where F: FnMut(&V) -> bool,

Returns an iterator over subslices separated by elements that match pred limited to returning at most n items. This starts at the end of the slice and works backwards. The matched element is not contained in the subslices.

See slice::rsplitn_mut for more details.

pub fn contains(&self, x: &V) -> boolwhere V: PartialEq,

Returns true if the slice contains an element with the given value.

See slice::contains for more details.

pub fn starts_with(&self, needle: &Self) -> boolwhere V: PartialEq,

Returns true if needle is a prefix of the slice.

See slice::starts_with for more details.

pub fn ends_with(&self, needle: &Self) -> boolwhere V: PartialEq,

Returns true if needle is a suffix of the slice.

See slice::ends_with for more details.

pub fn binary_search(&self, x: &V) -> Result<K, K>where V: Ord, K: From<usize>,

Binary searches this sorted slice for a given element.

See slice::binary_search for more details.

pub fn binary_search_by<'a, F>(&'a self, f: F) -> Result<K, K>where F: FnMut(&'a V) -> Ordering, K: From<usize>,

Binary searches this sorted slice with a comparator function.

See slice::binary_search_by for more details.

pub fn binary_search_by_key<'a, B, F>(&'a self, b: &B, f: F) -> Result<K, K>where F: FnMut(&'a V) -> B, B: Ord, K: From<usize>,

Binary searches this sorted slice with a key extraction function.

See slice::binary_search_by_key for more details.

pub fn sort_unstable(&mut self)where V: Ord,

Sorts the slice, but may not preserve the order of equal elements.

See slice::sort_unstable for more details.

pub fn sort_unstable_by<F>(&mut self, compare: F)where F: FnMut(&V, &V) -> Ordering,

Sorts the slice with a comparator function, but may not preserve the order of equal elements.

See slice::sort_unstable_by for more details.

pub fn sort_unstable_by_key<K2, F>(&mut self, f: F)where F: FnMut(&V) -> K2, K2: Ord,

Sorts the slice with a key extraction function, but may not preserve the order of equal elements.

See slice::sort_unstable_by_key for more details.

pub fn rotate_left(&mut self, mid: K)where usize: From<K>,

Rotates the slice in-place such that the first mid elements of the slice move to the end while the last self.next_key() - mid elements move to the front. After calling rotate_left, the element previously at index mid will become the first element in the slice.

See slice::rotate_left for more details.

pub fn rotate_right(&mut self, k: K)where usize: From<K>,

Rotates the slice in-place such that the first self.next_key() - k elements of the slice move to the end while the last k elements move to the front. After calling rotate_right, the element previously at index self.next_key() - k will become the first element in the slice.

See slice::rotate_right for more details.

pub fn clone_from_slice(&mut self, src: &Self)where V: Clone,

Copies the elements from src into self.

See slice::clone_from_slice for more details.

pub fn copy_from_slice(&mut self, src: &Self)where V: Copy,

Copies all elements from src into self, using a memcpy.

See slice::copy_from_slice for more details.

pub fn copy_within<R>(&mut self, src: R, dest: K)where R: TiRangeBounds<K>, V: Copy, usize: From<K>,

Copies elements from one part of the slice to another part of itself, using a memmove.

See slice::copy_within for more details.

pub fn swap_with_slice(&mut self, other: &mut Self)

Swaps all elements in self with those in other.

See slice::swap_with_slice for more details.

pub unsafe fn align_to<U>(&self) -> (&Self, &TiSlice<K, U>, &Self)

Transmute the slice to a slice of another type, ensuring alignment of the types is maintained.

See slice::align_to for more details.

pub unsafe fn align_to_mut<U>( &mut self ) -> (&mut Self, &mut TiSlice<K, U>, &mut Self)

Transmute the slice to a slice of another type, ensuring alignment of the types is maintained.

See slice::align_to_mut for more details.

impl<K> TiSlice<K, u8>

pub fn is_ascii(&self) -> bool

Checks if all bytes in this slice are within the ASCII range.

See slice::is_ascii for more details.

pub fn eq_ignore_ascii_case(&self, other: &Self) -> bool

Checks that two slices are an ASCII case-insensitive match.

See slice::eq_ignore_ascii_case for more details.

pub fn make_ascii_uppercase(&mut self)

Converts this slice to its ASCII upper case equivalent in-place.

See slice::make_ascii_uppercase for more details.

pub fn make_ascii_lowercase(&mut self)

Converts this slice to its ASCII lower case equivalent in-place.

See slice::make_ascii_lowercase for more details.

impl<K, V> TiSlice<K, V>

pub fn sort(&mut self)where V: Ord,

Sorts the slice.

See slice::sort for more details.

pub fn sort_by<F>(&mut self, compare: F)where F: FnMut(&V, &V) -> Ordering,

Sorts the slice with a comparator function.

See slice::sort_by for more details.

pub fn sort_by_key<K2, F>(&mut self, f: F)where F: FnMut(&V) -> K2, K2: Ord,

Sorts the slice with a key extraction function.

See slice::sort_by_key for more details.

pub fn sort_by_cached_key<K2, F>(&mut self, f: F)where F: FnMut(&V) -> K2, K2: Ord,

Sorts the slice with a key extraction function.

See slice::sort_by_cached_key for more details.

pub fn to_vec(&self) -> TiVec<K, V>where V: Clone,

Copies self into a new TiVec.

See slice::to_vec for more details.

pub fn into_vec(self: Box<Self>) -> TiVec<K, V>

Converts self into a vector without clones or allocation.

See slice::into_vec for more details.

pub fn repeat(&self, n: usize) -> TiVec<K, V>where V: Copy,

Creates a vector by repeating a slice n times.

See slice::repeat for more details.

pub fn concat<Item: ?Sized>(&self) -> <Self as Concat<Item>>::Outputwhere Self: Concat<Item>,

Flattens a slice of T into a single value Self::Output.

See slice::concat for more details.

pub fn join<Separator>( &self, sep: Separator ) -> <Self as Join<Separator>>::Outputwhere Self: Join<Separator>,

Flattens a slice of T into a single value Self::Output, placing a given separator between each.

See slice::join for more details.

impl<K> TiSlice<K, u8>

pub fn to_ascii_uppercase(&self) -> TiVec<K, u8>

Returns a vector containing a copy of this slice where each byte is mapped to its ASCII upper case equivalent.

See slice::to_ascii_uppercase for more details.

pub fn to_ascii_lowercase(&self) -> TiVec<K, u8>

Returns a vector containing a copy of this slice where each byte is mapped to its ASCII lower case equivalent.

See slice::to_ascii_lowercase for more details.