Struct typed_index_collections::TiVec

#[repr(transparent)]
pub struct TiVec<K, V> {
    pub raw: Vec<V>,
    /* private fields */
}

A contiguous growable array type that only accepts keys of the type K.

TiVec<K, V> is a wrapper around Rust container type std::vec::Vec. The struct mirrors the stable API of Rust std::vec::Vec and forwards to it as much as possible.

TiVec<K, V> uses K instead of usize for element indices and require the index to implement From<usize> and Into<usize> traits. Their implementation can be easily done with derive_more crate and #[derive(From, Into)].

TiVec<K, V> can be converted to std::vec::Vec<V> and back using From and Into.

There are also zero-cost conversions available between references:

&std::vec::Vec<V> and &TiVec<K, V> with AsRef,

&mut std::vec::Vec<V> and &mut TiVec<K, V> with AsMut,

Added methods:

from_ref - Converts a &std::vec::Vec<V> into a &TiVec<K, V>.

from_mut - Converts a &mut std::vec::Vec<V> into a &mut TiVec<K, V>.

push_and_get_key - Appends an element to the back of a collection and returns its index of type K.

pop_key_value - Removes the last element from a vector and returns it with its index of type K, or None if the vector is empty.

drain_enumerated - Creates a draining iterator that removes the specified range in the vector and yields the current count and the removed items. It acts like self.drain(range).enumerate(), but instead of usize it returns index of type K.

into_iter_enumerated - Converts the vector into iterator over all key-value pairs with K used for iteration indices. It acts like self.into_iter().enumerate(), but use K instead of usize for iteration indices.

Example

use typed_index_collections::TiVec;
use derive_more::{From, Into};

#[derive(From, Into)]
struct FooId(usize);

let mut foos: TiVec<FooId, usize> = std::vec![10, 11, 13].into();
foos.insert(FooId(2), 12);
assert_eq!(foos[FooId(2)], 12);

Fields
raw: Vec<V>

Raw slice property

Implementations

impl<K, V> TiVec<K, V>

pub fn new() -> Self

Constructs a new, empty TiVec<K, V>.

See Vec::new for more details.

pub fn with_capacity(capacity: usize) -> Self

Constructs a new, empty TiVec<K, V> with the specified capacity.

See Vec::with_capacity for more details.

pub unsafe fn from_raw_parts( ptr: *mut V, length: usize, capacity: usize ) -> Self

Creates a TiVec<K, V> directly from the raw components of another vector.

See Vec::from_raw_parts for more details.

pub fn from_ref(raw: &Vec<V>) -> &Self

Converts a &std::vec::Vec<V> into a &TiVec<K, V>.

Vector reference is intentionally used in the argument instead of slice reference for conversion with no-op.

Example

pub struct Id(usize);
let vec: &TiVec<Id, usize> = TiVec::from_ref(&vec![1, 2, 4]);

pub fn from_mut(raw: &mut Vec<V>) -> &mut Self

Converts a [&mut std::vec::Vec<V>] into a &mut TiVec<K, V>.

Example

pub struct Id(usize);
let vec: &mut TiVec<Id, usize> = TiVec::from_mut(&mut vec![1, 2, 4]);

pub fn capacity(&self) -> usize

Returns the number of elements the vector can hold without reallocating.

See Vec::capacity for more details.

pub fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted in the given TiVec<K, V>. The collection may reserve more space to avoid frequent reallocations. After calling reserve, capacity will be greater than or equal to self.len() + additional. Does nothing if capacity is already sufficient.

See Vec::reserve for more details.

pub fn reserve_exact(&mut self, additional: usize)

Reserves the minimum capacity for exactly additional more elements to be inserted in the given TiVec<K, V>. After calling reserve_exact, capacity will be greater than or equal to self.len() + additional. Does nothing if the capacity is already sufficient.

See Vec::reserve_exact for more details.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the vector as much as possible.

See Vec::shrink_to_fit for more details.

pub fn into_boxed_slice(self) -> Box<TiSlice<K, V>>

Converts the vector into Box<TiSlice<K, V>>.

See Vec::into_boxed_slice for more details.

pub fn truncate(&mut self, len: usize)

Shortens the vector, keeping the first len elements and dropping the rest.

See Vec::truncate for more details.

pub fn as_slice(&self) -> &TiSlice<K, V>

Extracts a slice containing the entire vector.

See Vec::as_slice for more details.

pub fn as_mut_slice(&mut self) -> &mut TiSlice<K, V>

Extracts a mutable slice of the entire vector.

See Vec::as_mut_slice for more details.

pub fn as_ptr(&self) -> *const V

Returns a raw pointer to the vector’s buffer.

See Vec::as_ptr for more details.

pub fn as_mut_ptr(&mut self) -> *mut V

Returns an unsafe mutable pointer to the vector’s buffer.

See Vec::as_mut_ptr for more details.

pub unsafe fn set_len(&mut self, new_len: usize)

Forces the length of the vector to new_len.

See Vec::set_len for more details.

pub fn swap_remove(&mut self, index: K) -> Vwhere usize: From<K>,

Removes an element from the vector and returns it.

The removed element is replaced by the last element of the vector.

See Vec::swap_remove for more details.

pub fn insert(&mut self, index: K, element: V)where usize: From<K>,

Inserts an element at position index within the vector, shifting all elements after it to the right.

See Vec::insert for more details.

pub fn remove(&mut self, index: K) -> Vwhere usize: From<K>,

Removes and returns the element at position index within the vector, shifting all elements after it to the left.

See Vec::remove for more details.

pub fn retain<F>(&mut self, f: F)where F: FnMut(&V) -> bool,

Retains only the elements specified by the predicate.

See Vec::retain for more details.

pub fn dedup_by_key<F, K2>(&mut self, key: F)where F: FnMut(&mut V) -> K2, K2: PartialEq,

Removes all but the first of consecutive elements in the vector that resolve to the same key.

See Vec::dedup_by_key for more details.

pub fn dedup_by<F>(&mut self, same_bucket: F)where F: FnMut(&mut V, &mut V) -> bool,

Removes all but the first of consecutive elements in the vector satisfying a given equality relation.

See Vec::dedup_by for more details.

pub fn push(&mut self, value: V)

Appends an element to the back of a collection.

See Vec::push for more details.

pub fn push_and_get_key(&mut self, value: V) -> Kwhere K: From<usize>,

Appends an element to the back of a collection and returns its index of type K.

It acts like { vec.push(...); vec.last_key().unwrap() }, but is optimized better.

See Vec::push for more details.

Example

#[derive(Eq, Debug, From, Into, PartialEq)]
pub struct Id(usize);
let mut vec: TiVec<Id, usize> = vec![1, 2, 4].into();
assert_eq!(vec.push_and_get_key(8), Id(3));
assert_eq!(vec.push_and_get_key(16), Id(4));
assert_eq!(vec.push_and_get_key(32), Id(5));

pub fn pop(&mut self) -> Option<V>

Removes the last element from a vector and returns it, or None if it is empty.

See Vec::pop for more details.

pub fn pop_key_value(&mut self) -> Option<(K, V)>where K: From<usize>,

Removes the last element from a vector and returns it with its index of type K, or None if the vector is empty.

See Vec::pop for more details.

Example

#[derive(Eq, Debug, From, Into, PartialEq)]
pub struct Id(usize);
let mut vec: TiVec<Id, usize> = vec![1, 2, 4].into();
assert_eq!(vec.pop_key_value(), Some((Id(2), 4)));
assert_eq!(vec.pop_key_value(), Some((Id(1), 2)));
assert_eq!(vec.pop_key_value(), Some((Id(0), 1)));
assert_eq!(vec.pop_key_value(), None);

pub fn append(&mut self, other: &mut Self)

Moves all the elements of other into Self, leaving other empty.

See Vec::append for more details.

pub fn drain<R>(&mut self, range: R) -> Drain<'_, V>where R: TiRangeBounds<K>,

Creates a draining iterator that removes the specified range in the vector and yields the removed items.

See Vec::drain for more details.

pub fn drain_enumerated<R>( &mut self, range: R ) -> TiEnumerated<Drain<'_, V>, K, V>where K: From<usize>, R: TiRangeBounds<K>,

Creates a draining iterator that removes the specified range in the vector and yields the current count and the removed items.

It acts like self.drain(range).enumerate(), but instead of usize it returns index of type K.

Note that the indices started from K::from_usize(0), regardless of the range starting point.

See Vec::drain for more details.

Example

#[derive(Eq, Debug, From, Into, PartialEq)]
pub struct Id(usize);
let mut vec: TiVec<Id, usize> = vec![1, 2, 4].into();
{
    let mut iterator = vec.drain_enumerated(Id(1)..);
    assert_eq!(iterator.next(), Some((Id(0), 2)));
    assert_eq!(iterator.next(), Some((Id(1), 4)));
    assert_eq!(iterator.next(), None);
}
assert_eq!(vec.as_slice(), TiSlice::from_ref(&[1]));

pub fn clear(&mut self)

Clears the vector, removing all values.

See Vec::clear for more details.

pub fn len(&self) -> usize

Returns the number of elements in the vector, also referred to as its ‘length’.

See Vec::len for more details.

pub fn is_empty(&self) -> bool

Returns true if the vector contains no elements.

See Vec::is_empty for more details.

pub fn split_off(&mut self, at: K) -> Selfwhere usize: From<K>,

Splits the collection into two at the given index.

See Vec::split_off for more details.

pub fn resize_with<F>(&mut self, new_len: usize, f: F)where F: FnMut() -> V,

Resizes the TiVec in-place so that len is equal to new_len.

See Vec::resize_with for more details.

pub fn resize(&mut self, new_len: usize, value: V)where V: Clone,

Resizes the TiVec in-place so that len is equal to new_len.

See Vec::resize for more details.

pub fn extend_from_slice(&mut self, other: &TiSlice<K, V>)where V: Clone,

Clones and appends all elements in a slice to the TiVec.

See Vec::extend_from_slice for more details.

pub fn dedup(&mut self)where V: PartialEq,

Removes consecutive repeated elements in the vector according to the PartialEq trait implementation.

See Vec::dedup for more details.

pub fn splice<R, I>( &mut self, range: R, replace_with: I ) -> Splice<'_, I::IntoIter>where R: TiRangeBounds<K>, I: IntoIterator<Item = V>,

Creates a splicing iterator that replaces the specified range in the vector with the given replace_with iterator and yields the removed items. replace_with does not need to be the same length as range.

See Vec::splice for more details.

pub fn into_iter_enumerated(self) -> TiEnumerated<IntoIter<V>, K, V>where K: From<usize>,

Converts the vector into iterator over all key-value pairs with K used for iteration indices.

It acts like self.into_iter().enumerate(), but use K instead of usize for iteration indices.

Example

#[derive(Eq, Debug, From, Into, PartialEq)]
pub struct Id(usize);
let vec: TiVec<Id, usize> = vec![1, 2, 4].into();
let mut iterator = vec.into_iter_enumerated();
assert_eq!(iterator.next(), Some((Id(0), 1)));
assert_eq!(iterator.next(), Some((Id(1), 2)));
assert_eq!(iterator.next(), Some((Id(2), 4)));
assert_eq!(iterator.next(), None);